Added setup scripts and fix CodeQL security warnings. Created PowerShell and Bash scripts to help users download JCuda dependencies automatically, making the initial setup much easier. Updated README with a Quick Start guide and troubleshooting section. Also fixed 6 CodeQL warnings including SLF4J format string issues, unused method parameters, and potential integer overflow in memory allocation calculations.

+++ src/test/java/com/vindex/jvectorcuda/EuclideanDistanceTest.java
@@ -177,9 +177,9 @@ class EuclideanDistanceTest {
         CUdeviceptr d_query = new CUdeviceptr();
         CUdeviceptr d_distances = new CUdeviceptr();
 
-        int dbSize = numVectors * dimensions * Sizeof.FLOAT;
+        long dbSize = (long) numVectors * dimensions * Sizeof.FLOAT;
         int querySize = dimensions * Sizeof.FLOAT;
-        int distSize = numVectors * Sizeof.FLOAT;
+        long distSize = (long) numVectors * Sizeof.FLOAT;
 
         cuMemAlloc(d_database, dbSize);
         cuMemAlloc(d_query, querySize);
